remove auth controller
